Consumido o serviço de sorteio e definido as mensagens de alerta

+++ frontend/src/redux/person/person.actions.js
@@ -1,3 +1,7 @@
+import React from 'react';
+import { toastr } from 'react-redux-toastr';
+import ConfirmAlert from '../../components/confirm-alert/cofirm-alert.component';
+
 import PersonTypes from './person.types';
 
 import api from '../../services/api';
@@ -6,10 +10,7 @@ export const fetchPeople = () => {
     return dispatchEvent => {
         api.get('people')
             .then(res => {
-                dispatchEvent({ type: PersonTypes.FETCH_PEOPLE, payload: res.data })
-            })
-            .catch(error => {
-                console.log(error);
+                dispatchEvent({ type: PersonTypes.FETCH_PEOPLE, payload: res.data });
             })
     }
 }
@@ -18,10 +19,11 @@ export const createPerson = person => {
     return dispatchEvent => {
         api.post('people', person)
             .then(res => {
-                dispatchEvent(fetchPeople())
+                dispatchEvent(fetchPeople());
+                toastr.success('Sucesso', 'Pessoa cadastrada!');
             })
             .catch(error => {
-                console.log(error)
+                toastr.error('Atenção', 'Problema ao cadastrar!');
             })
     }
 }
@@ -30,23 +32,33 @@ export const updatePerson = ({ _id, name, email }) => {
     return dispatchEvent => {
         api.put(`people/${_id}`, { name, email })
             .then(res => {
-                dispatchEvent(fetchPeople())
+                dispatchEvent(fetchPeople());
+                toastr.success('Sucesso', 'Pessoa atualizada!');
             })
             .catch(error => {
-                console.log(error)
+                toastr.error('Atenção', 'Problema ao atualizar!');
             })
     }
 }
 
 export const deletePerson = id => {
     return dispatchEvent => {
-        api.delete(`people/${id}`)
-            .then(res => {
-                dispatchEvent(fetchPeople())
-            })
-            .catch(error => {
-                console.log(error)
-            })
+        toastr.confirm(null, {
+            onOk: () => {
+                api.delete(`people/${id}`)
+                    .then(res => {
+                        dispatchEvent(fetchPeople())
+                        toastr.success('Sucesso', 'Pessoa deletada!');
+                    })
+                    .catch(error => {
+                        toastr.error('Atenção', 'Problema ao deletar!');
+                    })
+            },
+            onCancel: () => {},
+            component: () => (
+                <ConfirmAlert message='Você tem certeza disso?' />
+            )
+        })
     }
 }
 
@@ -54,10 +66,10 @@ export const drawPeople = () => {
     return dispatchEvent => {
         api.put('draws', {})
             .then(res => {
-                console.log(res)
+                toastr.success('Sucesso', 'Sorteio realizado e e-mails enviados!');
             })
             .catch(error => {
-                console.log(error)
+                toastr.error('Atenção', 'Problema ao realizar sorteio!');
             })
     }
 }
